Last minute update
Realized that the projectiles to the Wraithverge and Bloodscourge were not set to be non-reflective. Even though the Bloodscourge projectile works, the same cannot be said for the Wraithverge spirits. No matter what, if you fire the weapon at Heresiarch and he casts his invulnerability spell, those spirits will be coming after you. Not sure how to resolve this but this issue can be resolved in the next major update. I'll probably have to reprogram the projectiles so they cannot be reflected back at the player. I'll have to play around with some ideas.